Add Safir Kafka configurations and initializers
This adds the safir.events initializers for Kafka:

- configure_kafka_ssl
- init_recordname_schema_manager
- init_kafka_producer

It also adds the required configurations. With Pydantic we can do quite
a bit of validation of these configurations, specifically using enum for
Kafka protocol and schema compatibility settings, as well as validating
that the cert configurations are set if the SSL protocol is being used.