Add Kafka consumer
The consume_events function is run as a asyncio task from the
application so it doesn't block the http handlers.

The role of consume events is to run the AIOKafkaConsumer and
deserialize messages.

The route_messages function then creates tasks specifically for each
type of message that comes in.

New configurations:

- kafka_consumer_group_id