Add pytest fixtures to produce messages in tests
The purpose of these fixtures is to be able to produce Kafka messages
from pytest functions that the app can consume and run on.

The async_test_until function lets us poll that a condition is true,
within a timeout period. This is useful in async apps because we want to
test that certain log messages are sent, for example, but we may need to
wait for that to happen.