Route an edition.updated event from ltd.events
This adds routing for an edition.updated event from the ltd.events Kafka
topic. The processing is in a spawn aiojobs job.

To test this, we've set up a simulated LTD Events-like producer that
produces an edition.updated Kafka message. Since we don't have full
visibility inside the app, we capture logs, and test that the log
message from process_edition_updated() gets produced.