Classify content type from LTD product
This adds a new module, classification, that provides tools to classify
content sources using simple heuristics. Once content is classified, we
can plan how to ingest it.

The first workflow is classify_ltd_site, which presupposes that the
content is a website hosted on LSST the Docs.

Right now the heuristics are _very_ simple: is it a document or a
general website? Is it a Lander-based project with a metadata.jsonld
file? Or otherwise it's a Sphinx based technote. We can continue to
develop these heuristics as the need arises.

Each element is tested, as well as an overall real-life test of
classify_ltd_site against real sites on LSST the Docs.

Finally, this classification code is integrated into the
process_edition_updated event handler. The test for that now uses a
realistic event payload featuring the pipelines.lsst.io site.