Add queue_ltd_document_ingest
This adds the ability to queue an LTD document into the ingest Kafka
topic. All content ingest requests will go into this queue, but this
function specifically serializes the data with an ool.ltd_url_ingest_v1
schema that provides context about the LTD Product and Edition resources
that the URL is hosted in. For a "document", which is a single URL, the
ingested URL is the same as the edition's published URL.

The ook.url_key_v1 schema encodes the URL being ingested. The idea
behind that is the ingest topic can be partitioned, but all ingest
requests for a given URL will go to the same partition. This avoids
out-of-order ingest requests for the same URL.

To test this, test_process_edition_updated now specifically uses the
main edition of sqr-000.lsst.io, a Sphinx-based technote, as its test
case. This lets us look for the log message from the
queue_ltd_document_ingest function.