Create ReducedLtdSphinxTechnote
This commit puts in place all the infrastructure to take a technote's
HTML and it's metadata.yaml file, and reduce it. The
ReducedLtdSphinxTechnote class will become the basis of individual
records for Algolia (coming up next).

The sphinxutils module is based on work in
https://github.com/astropy/learn-astropy-librarian.

The idea of iter_sphinx_sections is to break down a sphinx document's
content into sections. These sections are small enough to create
relevant full-text content for Algolia search.

SQR-035 is used for testing the content reduction.

New dependencies:

- lxml
- cssselect
- pyyaml