Add ReducedLtdLanderDocument
This is a reducer for Lander-based documents (PDFs) on LSST the Docs.

With Lander projects, we pre-make a metadata.jsonld file that includes
extracted metadata and fulltext content. Thus we just have to parse the
plain text content made by pandoc into chunks (paragraph-sized pieces of
content that are associated with headers).