Implement object expiration by surrogate key
Some objects may not be replaced during an ingest (because their header
no longer exists, or their paragraph ID no longer exists, etc.). How we
deal with that is by setting a surrogateKey unique to each update for a
baseUrl.

Then to expire these old records, we search for records that do no
posess the *current* surrogateKey, and delete those in bulk.