Support Lander docs without content
Many LaTeX/Lander technotes have an abstract, but no content (for
whatever reason). To give these documents visibility (so that their
authors can eventually deal with them) we still want to ingest them.

To do that, if there are no content chunks we instead create a content
chunk using the abstract (normally the abstract doesn't have its own
chunk as its embedded with the rest of the document).