Handle AASTeX technotes w/o Lander support
AASTeX formatted technotes don't have the full support of Lander to
parse their content to get document handles or abstracts.

An example of that is DMTN-145.

To work around that, we're introducing a new Handle class that is
capable of parsing the handle from lsst.io URLs. This Handle class also
simplifies a lot of the fallback logic.