Add heuristic rejecting tex pandoc lets through
If pandoc fails to convert a TeX doc ot plaintext, it'll leave latex
behind. A telltale sign of TeX is that the first character is a
backslash! On that account, we can reject content that won't be useful
for search anyways (the best place to fix this is upstream, in the raw
content extraction.

This uses DMTN-119 as an example.