Create the GitHubMetadataService
The purpose of this service is to get metadata about a GitHub repository
as needed, often as a sub-service for document parsers.

This replaces the old github and utils modules, as well as the
githubcreationdate workflow (which is now the seed for this service).

We've optimized the algorithm for getting the date of the first commit
as well. No longer are we ignoring additional pages of content.