Add KafkaProducerDependency to the Factory
This makes it easier to spin up a Kafka producer from either HTTP path
operations or in the handlers for the Kafka consumer.

Now the Kafka consumer gets it's PydanticSchemaManager from the factory
rather than creating an entirely new one.

This mocking and kafka producer modules are based on Squarebot's; these
need to be moved up into Kafkit.