Add dependency for the Algolia SearchClient
The Algolia SearchClient manages an aiohttp client session. Therefore it
needs to be shut down.

It's a context manager, which is tricky to use as a dependency;
especially with our Factory pattern when it's run outside a FastAPI
response handler context. The way we work around this is to manually
call the __aenter___ and __aexit methods in the initialization and close
phases of the dependency.