Mock the Algolia SearchClient for testing
This is a minimal mock right now; more can be implemented later for
testing the usage of the search client.

Since the SearchClient is mocked, it means that the Algolia credentials
can always be required since the test environment supplies substitute
credentials. This will simplify the creation of a "ProcessContext" that
holds all reusable clients.