Refactor Factory with a ProcessContext
The purpose of this refactor is to make it easier to get a factory in
circumstances other than the FastAPI request handler (such as in a CLI
handler). The ProcessContext holds all clients that can persist across
requests. This design also nicely refactors the code for initializing
FastAPI dependencies out of the main module and into the
ProcessContext.create class method.

This is mimicking the structure of Gafaelfawr's Factory and
ProcessContext.