Stop using http_client_dep... in ProcessContext
I'm not sure why, but when creating a ProcessContext for a standalone
factory for service testing I found that the http client created by the
http_client_dependency was closed. Creating the http client insitu of
the ProcessContext solves this problem. I'm not sure why, and could be
worth more investigation.

Also fixed a bug where the kafka producer, schema manager, and algolia
client were being retrieved from dependencies to close them rather than
referencing the attributes already in ProcessContext.