Run Testcontainers Kafka via nox
This changes the approach previously taken where I ran testcontainers in
the pytest conftest.py module. Here I'm running the testcontainers in
the nox sessions, which is more similar to how tox-docker runs. The
advantage of this approach is that it doesn't require any code
modifications and patching because the environment variables for the
Kafka configuration can be made *before* starting up the app tests /
development run or documentation building sessions.

Running testcontainers via pytest could still be a useful technique if
different tests need full isolation of their testcontainers, but of
course this is a slow and resource-intensive approach.