Create a GitHub repo storage service
This is intended to be a general interface for getting metadata and
contents from a GitHub repository. The code for getting the recursive
contents tree and Git blobs is adapted from Times Square, and the code
for getting releases is new. This client could likely be upstreamed into
Safir because Ook, Times Square, Semaphore, and more apps all use GitHub
as a content source.