Switch to Factory for SdmSchemasIngestService
Switch to using the standalone Factory for creating the
SdmSchemasIngestService in the service's tests. The advantage of doing
so it is centralizes the concerns over the construction to the Factory
rather than needing to replicate the construction in individual tests.
This will be useful when we add SQL stores to the service.

To do this we needed to build out a generic mock for getting a GitHub
app client.