Create the SdmSchemasLinkStore
This datastore interfaces to the Postgres tables for SDM Schemas link
data. We've implemented an update_schema method that performs a bulk
"upsert" of the schema data. See
https://docs.sqlalchemy.org/en/20/dialects/postgresql.html#postgresql-insert-on-conflict

To make the upsert work safely we've added unique constraints on schema
entitiy names to ensure we can't add duplicate records for a given
entity.