Create a LinksService
This service in Ook will power all API requests to find links. Right now
it includes SDM links as a specific domain, but it'll grow to service
link requests from other domains like Python API documentation, etc.

Update the SdmSchemasLinkStore to provide a get_schema method.

Add a domain model for describing a link.

Note that we really should try to sync up this domain model with the
ingest models, if possible, and possible store the "kind" and
"source_title" info in the database rather than hard coding this at the
end.