Add Alembic integration and initial migration
This is a new way of orchestrating the alembic migrations by using nox
to run postgres, and using a sql dump to capture the existing schema
state of the databsae (as opposed to starting a database under an old
app and then switching branches and starting the database under the new
app).

Use .gitattributes to prevent the sql_dump from being code reviewed.