Refactor entity listing response model for links
For endpoints where we're returning a list of entities and their links,
I'm switching to a new SdmLink base type that has two fields: entity
info, and a list of the Links.

The entity info's type dependds on the exact nature of the entity, but
generally identifies the domain, domain type, the name of the
schema/table/column as appropriate, and finally the link to that
entity's links API resource.