Refactor LinkStore queries to use json agg
The goal of this work is to improve the queries made by Linkstore so
that each method uses only a single database query to do its work. This
will enable us to add pagination to the API in the future.

As well, the queries are now read directly into the domain models. This
is made possible by shaping the select to match the attributes of the
Pydantic domain models. Again, this is for future pagination needs.

Since the collections have a links attribute that is itself a list of
links, we're using json_agg in the queries to shape the query result
into embedded lists of dicts that are the same shape as the Pydantic
models.