Rename "name" to type specific names
Before SdmSchemaLlink, SdmTableLink, and SdmColumnLink domain models all
had a "name" attribute that was the type-specific name. However, this
makes it a little tricker to automatically create these link objects
from inputs because the name attribute means different things. Now
schema_name is always the schema's name across all types, rather than
being "name" for SdmSchemaLink in particular. We'll try this to see if
it's easier to use.