Run multi-type link queries with single statement
Now the LinkStore queries that return links from multiple types of SDM
entities run as a single SQL query. This is done in preparation for
pagination.

This affects get_sdm_links_scoped_to_schema (table and column links) and
get_sdm_links (schema, table, and column links)

To do this I had to use the new-to-me method of writing common table
expressions (CTE) for selecting from schema, table, and column links.

To feed the results directly into a pydantic model with its
model_validate method I had to introduce a Pydantic RootModel that is a
union of SdmSchemalinksCollection, SdmTableLinksCollection, and
SdmColumnLinksCollection. I use the entity type literal as the model
discriminator. A wrinkle in this approach is that now the underlying
domain models are available from the root attribute of the parsed
SdmColumnLinksCollection. This affects the response model transformation
step in the endpoint handler.