Implement pagination for SdmLinksCollections
This pagination for the endpoints that get combinations of schema,
table, and column links:

/limks/domains/sdm/schemas
/links/domains/sdm/schema/:schema/tables?include_columns=true

For these endpoints the SdmLinksCollectionCursor just sorts
alphabetically, for simplicity, by schema, table, and column names.
We've skipping consideration of the tap indices in these multi-type
collections because the cursor logic was becoming too complex.

We've also been able to refactor the code for building the select
statements so it's reused by multiple variants of the queries.