Add GitHubRepoStore.get_file_contents
This method uses the GitHub repository contents API to directly retrieve
a file based on its path. This is an alternative to getting the tree if
only a specific file directory or directory is required, whereas the
tree is useful for large-scale globbing.

Test with authordb.yaml data from lsst-texmf.