Implement authordb.yaml ingest
Add a new endpoint, POST /ook/ingest/lsst-texmf, that performs an ingest
of data in the lsst/lsst-texmf repo, particularly authordb.yaml right
now.

This is backed by the LsstTexmfIngestServide, which downloads the
authordb.yaml file and uses AuthorStore to upsert the affiliation,
collaboration, and author datasets into the database.

Includes a smoke test of this ingest using pre-downloaded datasets
registered through the GitHubMocker.