Add /authord endpoints by internal_id
This adds a new /authors router, and specifically two endpoints for
getting author information by ID:

- GET /authors/id/{author_id}
- GET /authors/project-access/id/{author_id}

The second endpoint includes the email address and is intended to be put
behind authentication.

The AuthorStore's method for doing the database query retireves
affiliation information with the author in a single query for
efficiency.