Install postgres extensions in testcontainers
This provides a simple way to ensure that extensions like pg_trgm are
installed in the postgres testcontainer on start-up that's independent
of runnng ook init, etc.

This requires us to add sqlalchemy and a postgres driver to the nox
testing dependencies.

The sleep before installing extensions seems necessary for reliability.
We might be able to get away with a shorter wait.