Provide option to delete stale authors
The default behaviour is to only sync authors that appear in
authordb.yaml, which results in authors previously ingested but no
longer in authordb.yaml remaining in the database. This new option
allows us to easily clean up the database, although in general use it
may be safer to not delete these stale records automatically.