Create domain models for Resources
These are the initial set of domain models for the Resources
(bibliographic) API. The idea is that these resources can map to
DataCite's schema for ease of creating DOIs.

The Resource class has specialized subclasses for adding additional
metadata for certain types of resources. For example, documents. These
will get modeled with Joined-Table Inheritance in the database.

To reference resources beyond the scope of Ook, we have the
ExternalReference domain model. This maps to relatedIdentifier or
relatedItem in the DataCite API. This domain model captures enough
detail to be able to make the relationship in DOI metadata.

A notable difference from our internal resources is that these resources
have simpler author information because we don't use the author db
tables.

Contributors, relations, and external references are all nullable so
that we don't need to load them for REST API queries.

model_rebuild is needed to resolve forward references in the pydantic
models.