Handle sphinx technotes with no abstract directive
Some Sphinx technote authors are leaving out the built-in abstract
directive, which means that the summary metadata isn't being populated.
In these cases, rather than fail the ingest just provide "Abstract not
available" as the summary. This will raise a conversation about the need
to use the built-in abstract directive.