Simplify querying for a resource
The goal here is to shape the data in the resource query so it matches
the domain models more closely. The new _queryutils module generates the
select statement with json aggregations that produce arrays of objects
to match related resources and contributor domain models, similar to
what's already done for the authorstore.

Also drop the ResourceLoadOptions concept because it complicates the
query at the moment, and the main goal is just to get back the domain
models.