Create a summary of related resources in responses
This adds the ability to return a summary of related resources to a
requested resource. We don't return the full related resource at this
point both to avoid performance concerns and also to avoid recusion
issues with a related resource also being related to the parent
resource, etc.

Things still to address:

1. testing of external references
2. continue simplifying the domain models.  This adds a new related
   attribute to Resource, in addition to the existing related_resource
   and related reference. Perhaps there should be separate domain models
   for ingest as there are for API responses?