Add number column to DocumentResource for sorting
Document handles aren't inherently sortable, so we add a numberic
"number" column to the DocumentResource table to support ordered
retrieval of documents in a given series.

The combination of series and number has a uniqueness constraint.

At the moment, we're not automatically extracting the number from the
handle, so clients should set this when adding a new document.