Implement /authors?search= endpoint
Add a search query parameter to the /authors endpoint to enable
filtering authors with a search by name. These returned author models
include a "score" attribute with a ranking of how well the author
matches the search term.

Author name search uses multiple strategies in order to handle different
input scenarios. Specifically we cover:

- "Last, First"
- "Last, Initial"
- "First Last"
- Family name only
- Given name only
- Compound family names
- Names with suffixes
- Partial names, initials, and typos

The first two formats are useful for handling author names as displayed in
publications. We use three different matching strategies:

- Trigram strategy with the pg_trgm extenion allows for flexible
  matching with typo tolerance,
- Component strategy handles matching name components as identified
  heuristically by the NameParser
- InitialStrategy handles matching when initials are known.

The SearchStrategy base class is a factory for generating the author
search query based on what components are identified in the name by the
NameParser.